requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
paho-mqtt==2.1.0
    # via -r requirements.in
